Added try catch block
