+++ src/test/java/com/cognite/client/ThreeDOutputsIntegrationTest.java
@@ -1,22 +1,20 @@
 package com.cognite.client;
 
-import com.cognite.client.dto.ThreeDModel;
-import com.cognite.client.dto.ThreeDModelRevision;
-import com.cognite.client.dto.ThreeDRevisionLog;
+
+import com.cognite.client.dto.*;
 import org.junit.jupiter.api.Tag;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.TestInstance;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import java.net.MalformedURLException;
 import java.time.Duration;
 import java.time.Instant;
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Map;
+import java.util.*;
 
 @TestInstance(TestInstance.Lifecycle.PER_CLASS)
-public class ThreeDRevisionLogTest extends ThreeDBaseTest {
+public class ThreeDOutputsIntegrationTest extends ThreeDBaseTest{
 
     private final Logger LOG = LoggerFactory.getLogger(this.getClass());
 
@@ -27,36 +25,30 @@ public class ThreeDRevisionLogTest extends ThreeDBaseTest {
 
     @Test
     @Tag("remoteCDP")
-    void listThreeDRevisionLogs() throws Exception {
+    void retrieveThreeDOutputs() throws MalformedURLException {
         try {
             Instant startInstant = Instant.now();
-            String loggingPrefix = "UnitTest - listThreeDModelsRevisions() - ";
+            String loggingPrefix = "retrieveThreeDOutputs - ";
+            LOG.info(loggingPrefix + "Start retrieving 3D Available Outputs.");
 
-            LOG.info(loggingPrefix + "Start list 3D Revision Logs.");
-
-            List<ThreeDRevisionLog> listResultsLogs = new ArrayList<>();
+            List<ThreeDOutput> listResultsOutputs = new ArrayList<>();
             for (Map.Entry<ThreeDModel, List<ThreeDModelRevision>> entry : super.map3D.entrySet()) {
                 ThreeDModel model = entry.getKey();
                 for (ThreeDModelRevision revision : entry.getValue()) {
-                    List<ThreeDRevisionLog> listResults =
+                    List<ThreeDOutput> listResults =
                             client.threeD()
                                     .models()
                                     .revisions()
-                                    .revisionLogs()
+                                    .outputs()
                                     .retrieve(model.getId(), revision.getId());
-                    listResultsLogs.addAll(listResults);
+                    listResultsOutputs.addAll(listResults);
                 }
             }
-
-            LOG.info(loggingPrefix + "Finished list 3D Revision Logs. Duration : {}",
+            LOG.info(loggingPrefix + "Finished retrieving 3D Available Outputs. Duration : {}",
                     Duration.between(startInstant, Instant.now()));
         } catch (Exception e) {
             LOG.error(e.toString());
             e.printStackTrace();
         }
-
     }
-
-
-
 }
